requirements.txt
dnspython==2.4.2
mysql-connector-python==8.2.0
protobuf==4.21.12
pymongo==4.6.1
